Move getParams method to helper
