+++ workers/loc.api/helpers/prepare-response/helpers/index.js
@@ -6,6 +6,7 @@ const omitPrivateModelFields = require('./omit-private-model-fields')
 const getBfxApiMethodName = require('./get-bfx-api-method-name')
 const getSymbolsForFiltering = require('./get-symbols-for-filtering')
 const getSymbolParams = require('./get-symbol-params')
+const getParams = require('./get-params')
 
 module.exports = {
   getParamsMap,
@@ -13,5 +14,6 @@ module.exports = {
   omitPrivateModelFields,
   getBfxApiMethodName,
   getSymbolsForFiltering,
-  getSymbolParams
+  getSymbolParams,
+  getParams
 }
